+++ packages/cards-editor-middleware/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/cards-editor-middleware
 
+## 0.4.60
+
+### Patch Changes
+
+-   Updated dependencies [1ce7fe9]
+    -   @sap-ux/project-access@1.29.15
+
 ## 0.4.59
 
 ### Patch Changes