+++ packages/cf-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.2.17
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.28
+
 ## 0.2.16
 
 ### Patch Changes