+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.33.15
+
+### Patch Changes
+
+-   Updated dependencies [1ce7fe9]
+    -   @sap-ux/project-access@1.29.15
+    -   @sap-ux/fiori-annotation-api@0.4.18
+
 ## 0.33.14
 
 ### Patch Changes