+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/fiori-elements-writer
 
+## 2.1.31
+
+### Patch Changes
+
+-   @sap-ux/annotation-generator@0.3.18
+-   @sap-ux/cap-config-writer@0.9.19
+-   @sap-ux/fe-fpm-writer@0.33.15
+-   @sap-ux/fiori-generator-shared@0.9.8
+-   @sap-ux/odata-service-writer@0.26.10
+-   @sap-ux/ui5-application-writer@1.3.1
+-   @sap-ux/ui5-test-writer@0.6.1
+
 ## 2.1.30
 
 ### Patch Changes